Do not use ObjectFilter when not required
